Added testing functionality to check for use with and without the clause argument being true.
